dom/tests/mochitest/bugs/test_bug612267.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=612267
-->
<head>
  <title>Test for Bug 393974</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=612267">Mozilla Bug 612267</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

Window.prototype.test = 'PASS';
is(window.test, 'PASS', "setting Window.prototype affects window.__proto__");
is(test, 'PASS', "setting Window.prototype affects the inner window lookup for sure");

</script>
</pre>
</body>
</html>
